fix: ReservationFacadeV2 적용
