--- src/main/kotlin/io/hhplus/concert/application/reservation/dto/ReservationResultV2.kt
@@ -1,8 +1,0 @@
-package io.hhplus.concert.application.reservation.dto
-
-data class ReservationResultV2(
-    val token: String,
-    val type: String,
-    val scheduleId: Long,
-    val rank: Long
-)
